+++ async-queues/src/broadcast/async_broadcast.rs
@@ -1,8 +1,8 @@
+use super::Config;
 use crate::*;
 
+use ::async_broadcast::{Receiver, RecvError, Sender};
 use ratelimit::Ratelimiter;
-use tokio::sync::broadcast::error::RecvError;
-use tokio::sync::broadcast::{Receiver, Sender};
 
 use core::sync::atomic::Ordering;
 use std::sync::Arc;
@@ -10,14 +10,13 @@ use std::sync::Arc;
 pub fn run(config: Config) -> Result<(), Box<dyn std::error::Error>> {
     let ratelimiter = config.ratelimiter();
 
-    // note: tokio broadcast channels always have `overflow` behavior where
-    // lagging subscribers will see messages dropped
-    let (tx, _rx) = tokio::sync::broadcast::channel::<Message>(config.queue_depth());
+    let (tx, mut rx) = ::async_broadcast::broadcast(config.queue_depth());
+    rx.set_overflow(true);
 
     let runtime = config.runtime();
 
     for _ in 0..config.subscribers() {
-        runtime.spawn_subscriber(receiver(tx.subscribe()));
+        runtime.spawn_subscriber(receiver(rx.clone()));
     }
 
     for _ in 0..config.publishers() {
@@ -31,14 +30,14 @@ pub fn run(config: Config) -> Result<(), Box<dyn std::error::Error>> {
 
 pub async fn receiver(mut rx: Receiver<Message>) {
     while RUNNING.load(Ordering::Relaxed) {
-        match rx.recv().await {
+        match rx.recv_direct().await {
             Ok(message) => {
                 message.validate();
 
                 RECV.increment();
                 RECV_OK.increment();
             }
-            Err(RecvError::Lagged(count)) => {
+            Err(RecvError::Overflowed(count)) => {
                 RECV.increment();
                 RECV_OVERFLOW.increment();
 
@@ -67,8 +66,7 @@ pub async fn sender(config: Config, tx: Sender<Message>, ratelimiter: Arc<Option
 
         let message = Message::new(config.message_length());
 
-        // note: tokio broadcast channel send is not async for some reason
-        if tx.send(message).is_ok() {
+        if tx.broadcast_direct(message).await.is_ok() {
             SEND.increment();
             SEND_OK.increment();
