+++ async-queues/src/broadcast/splaycast.rs
@@ -1,9 +1,9 @@
+use super::Config;
 use crate::*;
 
-use postage::dispatch::{Receiver, Sender};
-use postage::prelude::Sink;
-use postage::prelude::Stream;
+use futures_lite::StreamExt;
 use ratelimit::Ratelimiter;
+use splaycast::{Receiver, Sender};
 
 use core::sync::atomic::Ordering;
 use std::sync::Arc;
@@ -11,12 +11,17 @@ use std::sync::Arc;
 pub fn run(config: Config) -> Result<(), Box<dyn std::error::Error>> {
     let ratelimiter = config.ratelimiter();
 
-    let (tx, rx) = postage::dispatch::channel(config.queue_depth());
-
     let runtime = config.runtime();
 
+    let (tx, engine, splaycast) = splaycast::channel(config.queue_depth());
+
+    {
+        let _guard = runtime.fanout_runtime().enter();
+        tokio::spawn(engine);
+    }
+
     for _ in 0..config.subscribers() {
-        runtime.spawn_subscriber(receiver(rx.clone()));
+        runtime.spawn_subscriber(receiver(splaycast.subscribe()));
     }
 
     for _ in 0..config.publishers() {
@@ -30,12 +35,19 @@ pub fn run(config: Config) -> Result<(), Box<dyn std::error::Error>> {
 
 pub async fn receiver(mut rx: Receiver<Message>) {
     while RUNNING.load(Ordering::Relaxed) {
-        match rx.recv().await {
-            Some(message) => {
+        match rx.next().await {
+            Some(splaycast::Message::Entry { item }) => {
+                let message = item;
                 message.validate();
 
                 RECV.increment();
                 RECV_OK.increment();
+            }
+            Some(splaycast::Message::Lagged { count }) => {
+                RECV.increment();
+                RECV_OVERFLOW.increment();
+
+                DROPPED.add(count.try_into().unwrap());
             }
             None => {
                 RECV.increment();
@@ -45,11 +57,7 @@ pub async fn receiver(mut rx: Receiver<Message>) {
     }
 }
 
-pub async fn sender(
-    config: Config,
-    mut tx: Sender<Message>,
-    ratelimiter: Arc<Option<Ratelimiter>>,
-) {
+pub async fn sender(config: Config, tx: Sender<Message>, ratelimiter: Arc<Option<Ratelimiter>>) {
     while !RUNNING.load(Ordering::Relaxed) {
         std::hint::spin_loop()
     }
@@ -64,7 +72,7 @@ pub async fn sender(
 
         let message = Message::new(config.message_length());
 
-        if tx.send(message).await.is_ok() {
+        if tx.send(message).is_ok() {
             SEND.increment();
             SEND_OK.increment();
 
@@ -72,8 +80,6 @@ pub async fn sender(
         } else {
             SEND.increment();
             SEND_EX.increment();
-
-            DROPPED.increment();
         }
     }
 }
