+++ async-queues/src/broadcast/widecast.rs
@@ -1,7 +1,8 @@
+use super::Config;
 use crate::*;
 
-use broadcaster::*;
 use ratelimit::Ratelimiter;
+use widecast::*;
 
 use core::sync::atomic::Ordering;
 use std::sync::Arc;
@@ -13,15 +14,13 @@ pub fn run(config: Config) -> Result<(), Box<dyn std::error::Error>> {
 
     // note: broadcaster's channels always have `overflow` behavior where
     // lagging subscribers will see messages dropped
-    let tx = broadcaster::channel::<Message>(
-        runtime.fanout_runtime(),
-        config.queue_depth(),
-        config.fanout(),
-    );
+    let tx = widecast::Sender::<Message>::new(config.queue_depth());
 
     for _ in 0..config.subscribers() {
         runtime.spawn_subscriber(receiver(tx.subscribe()));
     }
+
+    let tx = Arc::new(tokio::sync::Mutex::new(tx));
 
     for _ in 0..config.publishers() {
         runtime.spawn_publisher(sender(config.clone(), tx.clone(), ratelimiter.clone()));
@@ -55,7 +54,11 @@ pub async fn receiver(mut rx: Receiver<Message>) {
     }
 }
 
-pub async fn sender(config: Config, tx: Sender<Message>, ratelimiter: Arc<Option<Ratelimiter>>) {
+pub async fn sender(
+    config: Config,
+    tx: Arc<tokio::sync::Mutex<Sender<Message>>>,
+    ratelimiter: Arc<Option<Ratelimiter>>,
+) {
     while !RUNNING.load(Ordering::Relaxed) {
         std::hint::spin_loop()
     }
@@ -70,15 +73,11 @@ pub async fn sender(config: Config, tx: Sender<Message>, ratelimiter: Arc<Option
 
         let message = Message::new(config.message_length());
 
-        // note: tokio broadcast channel send is not async for some reason
-        if tx.send(message).is_ok() {
-            SEND.increment();
-            SEND_OK.increment();
+        let mut sender = tx.lock().await;
+        sender.send(message);
+        SEND.increment();
+        SEND_OK.increment();
 
-            SEND_BYTES.add(config.message_length() as u64);
-        } else {
-            SEND.increment();
-            SEND_EX.increment();
-        }
+        SEND_BYTES.add(config.message_length() as u64);
     }
 }
