+++ async-queues/src/mpmc/postage.rs
@@ -1,8 +1,10 @@
+use super::Config;
 use crate::*;
 
-use futures_lite::StreamExt;
+use postage::dispatch::{Receiver, Sender};
+use postage::prelude::Sink;
+use postage::prelude::Stream;
 use ratelimit::Ratelimiter;
-use splaycast::{Receiver, Sender};
 
 use core::sync::atomic::Ordering;
 use std::sync::Arc;
@@ -10,20 +12,15 @@ use std::sync::Arc;
 pub fn run(config: Config) -> Result<(), Box<dyn std::error::Error>> {
     let ratelimiter = config.ratelimiter();
 
+    let (tx, rx) = postage::dispatch::channel(config.queue_depth());
+
     let runtime = config.runtime();
 
-    let (tx, engine, splaycast) = splaycast::channel(config.queue_depth());
-
-    {
-        let _guard = runtime.fanout_runtime().enter();
-        tokio::spawn(engine);
+    for _ in 0..config.consumers() {
+        runtime.spawn_subscriber(receiver(rx.clone()));
     }
 
-    for _ in 0..config.subscribers() {
-        runtime.spawn_subscriber(receiver(splaycast.subscribe()));
-    }
-
-    for _ in 0..config.publishers() {
+    for _ in 0..config.producers() {
         runtime.spawn_publisher(sender(config.clone(), tx.clone(), ratelimiter.clone()));
     }
 
@@ -34,19 +31,12 @@ pub fn run(config: Config) -> Result<(), Box<dyn std::error::Error>> {
 
 pub async fn receiver(mut rx: Receiver<Message>) {
     while RUNNING.load(Ordering::Relaxed) {
-        match rx.next().await {
-            Some(splaycast::Message::Entry { item }) => {
-                let message = item;
+        match rx.recv().await {
+            Some(message) => {
                 message.validate();
 
                 RECV.increment();
                 RECV_OK.increment();
-            }
-            Some(splaycast::Message::Lagged { count }) => {
-                RECV.increment();
-                RECV_OVERFLOW.increment();
-
-                DROPPED.add(count.try_into().unwrap());
             }
             None => {
                 RECV.increment();
@@ -56,7 +46,11 @@ pub async fn receiver(mut rx: Receiver<Message>) {
     }
 }
 
-pub async fn sender(config: Config, tx: Sender<Message>, ratelimiter: Arc<Option<Ratelimiter>>) {
+pub async fn sender(
+    config: Config,
+    mut tx: Sender<Message>,
+    ratelimiter: Arc<Option<Ratelimiter>>,
+) {
     while !RUNNING.load(Ordering::Relaxed) {
         std::hint::spin_loop()
     }
@@ -71,7 +65,7 @@ pub async fn sender(config: Config, tx: Sender<Message>, ratelimiter: Arc<Option
 
         let message = Message::new(config.message_length());
 
-        if tx.send(message).is_ok() {
+        if tx.send(message).await.is_ok() {
             SEND.increment();
             SEND_OK.increment();
 
@@ -79,6 +73,8 @@ pub async fn sender(config: Config, tx: Sender<Message>, ratelimiter: Arc<Option
         } else {
             SEND.increment();
             SEND_EX.increment();
+
+            DROPPED.increment();
         }
     }
 }
